Additional work to catch up with the transition from int to long values for Word (formerly IntConst).  This includes using long rather than int values in NonZero objects to match changes in Word; using 1L<<n rather than 1<<n to calculate 2^n as a long; and also includes some continuing attempts to use appropriate range checks in external generators (but still not confident these are correct; further review needed).

+++ src/mil/Atom.java
@@ -289,9 +289,9 @@ public abstract class Atom {
 
   /**
    * Return the nonzero value associated with this atom; a return of zero indicates that the atom
-   * was not an NonZero.
+   * was not a NonZero.
    */
-  int getNonZero() {
+  long getNonZero() {
     return 0;
   }
